Created a new "schematic" package
The intention is to move the contents of schematic into a separate egg fairly soon. This commit is an initial step to separate the schema and model/document components of mongothon into distinct modules prior to this extraction, with the goal of eliminating any dependencies on Mongothon from schema code.

Details:
* There is now a Mongothon Schema class which subclasses the Schematic Schema class - adding some Mongo specific doc spec defaults
* Sample schemas used in tests have been duplicated for each package's tests. The intention is to customize these to appropriate domain and remove the dependency on PyMongo from schematic code.